CLEAN: Login_Comp new file
