+++ src/pages/database/section/loginUI/Login_Comp.jsx
@@ -1,0 +1,40 @@
+import { useRef } from "react";
+import PropTypes from 'prop-types';
+import { authentication } from '@/assets/database/firebase'
+
+Login.propTypes = {
+    loggedInState: PropTypes.shape({loggedIn: PropTypes.bool, setLoggedIn: PropTypes.func})
+};
+export default function Login ({loggedInState}){
+
+    let emailInputRef = useRef(null);
+    let passwordInputRef = useRef(null);
+
+    const onSubmitLoginHandler = (e) => {
+        e.preventDefault();
+        
+        if(!loggedInState.loggedIn){
+            const email = emailInputRef.current.value;
+            const password = passwordInputRef.current.value;
+
+            (async function(){
+                loggedInState.setLoggedIn(await authentication.login(email, password))
+            })()
+        }else{
+            console.error("Already Logged in, request was deny")
+        }
+    }
+    
+    return (
+        <>
+            <div id='login'>
+                <h1>Login</h1>
+                <form onSubmit={onSubmitLoginHandler}>
+                    <input ref={emailInputRef} type='email'/>
+                    <input ref={passwordInputRef} type='password'/>
+                    <button type='submit'>Login</button>
+                </form>
+            </div>
+        </>
+    )
+}
